fix(admin-ui): keep delegation routes export-safe
