+++ openbank-admin-ui/src/app/delegations/page.tsx
@@ -19,31 +19,19 @@
 'use client'
 
 import { useCallback, useEffect, useRef, useState } from 'react'
-import Link from 'next/link'
 import { Search, Share2, RefreshCw, Activity } from 'lucide-react'
 import { useLanguage } from '@/lib/i18n/LanguageContext'
 import { classifyBffFailure } from '@/lib/services/bff'
 import { DataUnavailable, type UnavailableKind } from '@/components/feedback/DataUnavailable'
-import { EntityChip } from '@/components/entities/EntityChip'
 import { PageHeader } from '@/components/ui/PageHeader'
 import { RoleCatalog } from '@/components/delegations/RoleCatalog'
 import {
   EffectiveAccess,
-  effectiveResourceDetails,
-  grantConditions,
-  grantResourcePresentation,
   isEffectiveAccessPayload,
-  matchedRoleName,
   type EffectiveAccessPayload,
 } from '@/components/delegations/EffectiveAccess'
-import { capabilityLabel, isAssignablePresetCapability } from '@/lib/delegations/rolePresets'
-import { LegacyCapabilityEvidence } from '@/components/delegations/LegacyCapabilityEvidence'
-import {
-  DelegationStatusBadge,
-  counterpartyLabel,
-  grantCounterparty,
-  type Grant,
-} from '@/components/delegations/GrantView'
+import { GrantTable } from '@/components/delegations/GrantTable'
+import type { Grant } from '@/components/delegations/GrantView'
 
 type EntityRef = { type: string; id: string; label: string; sublabel?: string }
 
@@ -331,76 +319,6 @@ export default function DelegationsPage() {
           />
           <ProjectionHealth consumers={consumers} known={projectionKnown} loading={projectionLoading} />
         </>
-      )}
-    </div>
-  )
-}
-
-export function GrantTable({
-  title, subtitle, grants, state, direction, effectiveAccess,
-}: { title: string; subtitle: string; grants: Grant[]; state: DirectionState; direction: 'granted' | 'received'; effectiveAccess: EffectiveAccessPayload | null }) {
-  const { t, language } = useLanguage()
-  const details = effectiveAccess ? effectiveResourceDetails(effectiveAccess) : []
-
-  return (
-    <div className="card" style={{ padding: '16px', marginBottom: '20px' }}>
-      <h2 style={{ fontSize: '15px', fontWeight: 700, marginBottom: '2px' }}>{title}</h2>
-      <p style={{ fontSize: '12px', color: 'var(--text-tertiary)', marginBottom: '12px' }}>{subtitle}</p>
-
-      {state !== 'ok' ? (
-        <div style={{ padding: '16px', borderRadius: '8px', background: 'var(--surface-3)', fontSize: '13px' }}>
-          {state === 'forbidden'
-            ? t(
-                'Tento pohled vám nebyl povolen — nezaměňujte za „žádné delegace“.',
-                'This view was refused for your role — do not read it as “no delegations”.',
-              )
-            : t(
-                'Tento pohled se teď nepodařilo načíst — nezaměňujte za „žádné delegace“.',
-                'This view could not be loaded right now — do not read it as “no delegations”.',
-              )}
-        </div>
-      ) : grants.length === 0 ? (
-        <div style={{ padding: '16px', color: 'var(--text-tertiary)', fontSize: '13px' }}>
-          {t('Žádné delegace.', 'No delegations.')}
-        </div>
-      ) : (
-        <div style={{ overflowX: 'auto' }}>
-          <table className="table" style={{ width: '100%' }}>
-            <thead>
-              <tr>
-                <th>{t('Stav', 'Status')}</th>
-                <th>{t('Protistrana', 'Counterparty')}</th>
-                <th>{t('Role', 'Role')}</th>
-                <th>{t('Zdroj', 'Resource')}</th>
-                <th>{t('Práva', 'Rights')}</th>
-                <th>{t('Podmínky', 'Conditions')}</th>
-                <th />
-              </tr>
-            </thead>
-            <tbody>
-              {grants.map(g => {
-                const counterparty = grantCounterparty(g, direction)
-                const resource = grantResourcePresentation(g, details, language)
-                const role = effectiveAccess?.sources.presets === 'ok'
-                  ? matchedRoleName(g, effectiveAccess.presets, language)
-                  : t('Role není dostupná', 'Role unavailable')
-                return <tr key={g.id}>
-                  <td><DelegationStatusBadge status={g.status} /></td>
-                  <td><EntityChip type="party" id={counterparty.id} label={counterparty.name} /></td>
-                  <td style={{ fontSize: '12px', fontWeight: 650 }}>{role}</td>
-                  <td style={{ fontSize: '12px' }}><strong style={{ display: 'block' }}>{resource.label}</strong>{resource.meta && <span style={{ color: 'var(--text-tertiary)' }}>{resource.meta}</span>}</td>
-                  <td><div aria-label={t('Účinná práva', 'Effective rights')} style={{ display: 'flex', flexWrap: 'wrap', gap: 4 }}>{g.capabilities.filter(isAssignablePresetCapability).map(capability => <span key={capability} title={capability} style={{ borderRadius: 999, padding: '3px 7px', fontSize: 10, background: 'var(--surface-3)', border: '1px solid var(--border)' }}>{capabilityLabel(capability, language)}</span>)}</div><LegacyCapabilityEvidence capabilities={g.capabilities} /></td>
-                  <td style={{ fontSize: '11px' }}>{grantConditions(g, language).map(condition => <div key={condition.label}><span style={{ color: 'var(--text-tertiary)' }}>{condition.label}:</span> <strong>{condition.value}</strong></div>)}</td>
-                  <td>
-                    <Link href={`/delegations/${g.id}`} className="btn btn-secondary" style={{ fontSize: '12px' }}>
-                      {t('Detail', 'Detail')}
-                    </Link>
-                  </td>
-                </tr>
-              })}
-            </tbody>
-          </table>
-        </div>
       )}
     </div>
   )
