+++ openbank-admin-ui/src/test/delegation-coverage-probe.test.tsx
@@ -3,8 +3,8 @@
 import { cleanup, fireEvent, render, screen, within } from '@testing-library/react'
 import { SessionProvider } from 'next-auth/react'
 import { afterEach, describe, expect, it } from 'vitest'
-import { CoverageProbe } from '@/app/delegations/[id]/page'
-import { GrantTable } from '@/app/delegations/page'
+import { CoverageProbe } from '@/components/delegations/CoverageProbe'
+import { GrantTable } from '@/components/delegations/GrantTable'
 import type { EffectiveAccessPayload } from '@/components/delegations/EffectiveAccess'
 import type { Grant } from '@/components/delegations/GrantView'
 import { LanguageProvider } from '@/lib/i18n/LanguageContext'
